discord/requirements.txt
# King's Choice Discord Bot Dependencies

# Core Discord Bot Framework
discord.py>=2.3.0
aiohttp>=3.8.0

# Database
aiosqlite>=0.19.0

# Translation
deep-translator>=1.11.0

# Environment and Configuration
python-dotenv>=1.0.0

# Logging and Utilities
colorama>=0.4.6

# Development Dependencies (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0